refactor: move sampling strategy interfaces to internal/sampling/strategy
