+++ internal/sampling/http/cfgmgr.go
@@ -6,13 +6,13 @@ package http
 import (
 	"context"
 
-	"github.com/jaegertracing/jaeger/cmd/collector/app/sampling/samplingstrategy"
+	"github.com/jaegertracing/jaeger/internal/sampling/strategy"
 	"github.com/jaegertracing/jaeger/proto-gen/api_v2"
 )
 
 // ConfigManager implements ClientConfigManager.
 type ConfigManager struct {
-	SamplingProvider samplingstrategy.Provider
+	SamplingProvider strategy.Provider
 }
 
 // GetSamplingStrategy implements ClientConfigManager.GetSamplingStrategy.
